Added a progress bar to Quote Maintenance to show which parameters we're waiting for
Added a header bar to quote panel. It matches layout with Address and Job list, but it also gives
me a place to put the edit icon for updates. (Spacing is off, but I'll fix that later)

Added a "can edit" flag to the quote detail response because the calculation about who-is-allowed-to-do-what
is a little more complicated than seems appropriate to do client-side.
	modified:   WebContent/quoteMaintenance.jsp
	modified:   src/com/ansi/scilla/web/quote/response/QuoteListResponse.java
	modified:   src/com/ansi/scilla/web/quote/response/QuoteResponseItem.java
	modified:   src/com/ansi/scilla/web/quote/servlet/QuoteServlet.java