Called new validation functions and got it to compile and run.
I added the call to execute the validation from the 'validateRows'
function within timesheetImportResponse.java

I added a private variable to stash the Hasmap/Map of messages in
I added a getter to get to that.

I runs and is able to still display the grid however..
I am getting a bunch of JDBC Connection errors
from within the threadPool validations going on.