Was able to get the employee code delivered to the employee modal
added an employee code sql lookup style lookup function
  to the getEmployeeCode method
It checks to see it employeeCode has been populated and if not
  looks it up and stores it in the private member variable.
Caveat : I didn't see a conn being passed in to use..
  so it currently creates a conntection to the dev db via utils..

Also changed the top row in the model to not be editable.

Next Up - making the coninue button re-validate the modal data

On PTO til the Jul 6.. won't be touching this til the 7th.