Special Special Overrides: * Unactivate a Job * Create Clearly Windows Tickets
"Special" because these didn't fit nicely into the "Put everythingin the
enum" paradigm. Added an AbstractOverrideServlet that can be extended by
these special needs functions. The override servlet recognizes that this
is a different thing and calls the doGet()/doPost() in the special
servlet instead of processing via enum. The same responses are returned
in all cases so the JSP does not need to change
Changes to be committed:
	modified:   src/com/ansi/scilla/web/common/response/WebMessages.java
	modified:   src/com/ansi/scilla/web/specialOverride/common/SpecialOverrideType.java
	modified:   src/com/ansi/scilla/web/specialOverride/response/NameDescriptionResponseItem.java
	modified:   src/com/ansi/scilla/web/specialOverride/response/SpecialOverrideResponse.java
	modified:   src/com/ansi/scilla/web/specialOverride/response/SpecialOverrideSelectItem.java
	new file:   src/com/ansi/scilla/web/specialOverride/servlet/AbstractOverrideServlet.java
	new file:   src/com/ansi/scilla/web/specialOverride/servlet/ClearlyWindowsTicketServlet.java
	modified:   src/com/ansi/scilla/web/specialOverride/servlet/SpecialOverrideServlet.java
	new file:   src/com/ansi/scilla/web/specialOverride/servlet/UnactivateJobServlet.java